Rework how streaming metrics are stored & fetched
